analysis/detailed_data_collection/canonical_url_collection/process_channel_url/requirements.txt
google-api-python-client
google-cloud-pubsub
google-cloud-bigquery
requests
numpy
pandas
db-dtypes
beautifulsoup4